+++ packages/react-cosmos-playground2/src/shared/ui/valueInputTree/ValueInputTreeItem/StringItem.tsx
@@ -1,52 +1,33 @@
 import React from 'react';
-import { KEY_DOWN, KEY_UP } from '../../../../shared/keys';
 import {
   Label,
+  ValueContainer,
+  TextInputContainer,
   TextContainer,
   TextField,
-  TextInputContainer,
-  TextMirror,
-  ValueContainer
+  TextMirror
 } from './shared';
 
 type Props = {
   id: string;
   label: string;
-  value: number;
-  onChange: (newValue: number) => unknown;
+  value: string;
+  onChange: (newValue: string) => unknown;
 };
 
-export function NumberInput({ id, label, value, onChange }: Props) {
+export function StringItem({ id, label, value, onChange }: Props) {
   const [focused, setFocused] = React.useState(false);
   const onFocus = React.useCallback(() => setFocused(true), []);
   const onBlur = React.useCallback(() => setFocused(false), []);
 
   const onInputChange = React.useCallback(
-    (e: React.SyntheticEvent<HTMLTextAreaElement>) => {
-      const newValue = +e.currentTarget.value;
-      if (isFinite(newValue)) {
-        onChange(newValue);
-      }
-    },
+    (e: React.SyntheticEvent<HTMLTextAreaElement>) =>
+      onChange(e.currentTarget.value),
     [onChange]
   );
 
-  const onKeyDown = React.useCallback(
-    (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
-      switch (e.keyCode) {
-        case KEY_UP:
-          e.preventDefault();
-          return onChange(value + 1);
-        case KEY_DOWN:
-          e.preventDefault();
-          return onChange(value - 1);
-        default:
-        // Nada
-      }
-    },
-    [value, onChange]
-  );
-
+  // Mirror textarea behavior and add an extra row after user adds a new line
+  const mirrorText = focused ? value.replace(/\n$/, `\n `) : value;
   return (
     <>
       <Label title={label} htmlFor={id}>
@@ -56,7 +37,7 @@ export function NumberInput({ id, label, value, onChange }: Props) {
         <TextInputContainer focused={focused}>
           <TextContainer>
             <TextMirror style={{ opacity: focused ? 0 : 1 }}>
-              {value}
+              {mirrorText}
             </TextMirror>
             <TextField
               rows={1}
@@ -65,7 +46,6 @@ export function NumberInput({ id, label, value, onChange }: Props) {
               onChange={onInputChange}
               onFocus={onFocus}
               onBlur={onBlur}
-              onKeyDown={onKeyDown}
               style={{ opacity: focused ? 1 : 0 }}
             />
           </TextContainer>
